+++ src/cli/commands/config/index.js
@@ -23,16 +23,17 @@ const OPTIONS = require('../../../options');
 
 const EPILOGUE = `Available configuration options:
 
-  ${'CLI'.underline} - Manages the Emulator.
+  ${'CLI'.underline} - Global settings.
     ${'projectId'.bold}
       ${OPTIONS.projectId.description}
 
     ${'region'.bold}
       ${OPTIONS.region.description}
 
-    ${'service'.bold}
-      ${OPTIONS.service.description}
+    ${'storage'.bold}
+      ${OPTIONS.storage.description}
 
+  ${'CLI'.underline} - Manages the Emulator.
     ${'timeout'.bold}
       ${OPTIONS.timeout.description}
 
@@ -40,14 +41,17 @@ const EPILOGUE = `Available configuration options:
     ${'bindHost'.bold}
       ${OPTIONS.bindHost.description}
 
-    ${'grpcPort'.bold}
-      ${OPTIONS.grpcPort.description}
-
     ${'host'.bold}
       ${OPTIONS.host.description}
 
+    ${'idlePruneInterval'.bold}
+      ${OPTIONS.idlePruneInterval.description}
+
     ${'logFile'.bold}
       ${OPTIONS.logFile.description}
+
+    ${'maxIdle'.bold}
+      ${OPTIONS.maxIdle.description}
 
     ${'restPort'.bold}
       ${OPTIONS.restPort.description}
@@ -55,14 +59,20 @@ const EPILOGUE = `Available configuration options:
     ${'supervisorPort'.bold}
       ${OPTIONS.supervisorPort.description}
 
+    ${'tail'.bold}
+      ${OPTIONS.tail.description}
+
     ${'useMocks'.bold}
       ${OPTIONS.useMocks.description}
 
     ${'verbose'.bold}
       ${OPTIONS.verbose.description}
 
-    ${'tail'.bold}
-      ${OPTIONS.tail.description}
+    ${'watch'.bold}
+      ${OPTIONS.watch.description}
+
+    ${'watchIgnore'.bold}
+      ${OPTIONS.watchIgnore.description}
 
     See ${'https://github.com/GoogleCloudPlatform/cloud-functions-emulator/wiki/Configuring-the-Emulator'.bold}
 `;