+++ src/model/index.js
@@ -21,4 +21,3 @@ exports.CloudFunction = require('./cloudfunction');
 exports.Functions = Functions;
 exports.functions = (...args) => new Functions(...args);
 exports.Operation = require('./operation');
-exports.protos = require('./protos');