+++ test/system/cli/index.test.js
@@ -45,7 +45,6 @@ const prefix = `Google Cloud Functions Emulator`;
 const GCLOUD = process.env.GCLOUD_CMD_OVERRIDE || `gcloud`;
 const HOST = 'localhost';
 const REST_PORT = 8088;
-const GRPC_PORT = 8089;
 const SUPERVISOR_PORT = 8090;
 const REGION = `us-central1`;
 const PROJECT_ID = detectProjectId(null, false);
@@ -87,14 +86,12 @@ function makeTests (service, override) {
           server.set('service', service);
           config.set('restPort', REST_PORT);
           server.set('restPort', REST_PORT);
-          config.set('grpcPort', GRPC_PORT);
-          server.set('grpcPort', GRPC_PORT);
           config.set('supervisorPort', SUPERVISOR_PORT);
           server.set('supervisorPort', SUPERVISOR_PORT);
           config.set('logFile', logFile);
           server.set('logFile', logFile);
 
-          return tools.spawnAsyncWithIO('node', ['bin/functions', 'kill'], cwd);
+          return tools.spawnAsyncWithIO('node', ['bin/functions', 'stop'], cwd);
         })
         .then(() => tools.spawnAsyncWithIO('node', ['bin/functions', 'restart'], cwd))
         .then((results) => {
@@ -141,7 +138,6 @@ function makeTests (service, override) {
       it(`should list configuration`, () => {
         let output = run(`${cmd} config list`, cwd);
         assert(output.includes(`host`));
-        assert(output.includes(`grpcPort`));
         assert(output.includes(`restPort`));
       });
     });
@@ -152,7 +148,7 @@ function makeTests (service, override) {
       before(() => {
         if (override) {
           deployArgs = `${overrideArgs} --stage-bucket=${bucketName}`;
-          const output = run(`${override} list --region=${REGION}`, cwd);
+          const output = run(`${override} list --regions=${REGION}`, cwd);
           assert(output.includes(`Listed 0 items.`));
         } else {
           return tools.tryTest(() => {
@@ -163,7 +159,7 @@ function makeTests (service, override) {
       });
 
       it(`should deploy a background function`, () => {
-        const output = run(`${override || cmd} deploy ${name} --local-path=test/test_module/ --trigger-bucket=test ${deployArgs}`, cwd);
+        const output = run(`${override || cmd} deploy ${name} --source=test/test_module/ --trigger-bucket=test ${deployArgs}`, cwd);
         if (override) {
           try {
             assert(output.includes(`/functions/${name}`));
@@ -177,7 +173,7 @@ function makeTests (service, override) {
       });
 
       it(`should deploy a function with JSON`, () => {
-        const output = run(`${override || cmd} deploy helloData --local-path=test/test_module/ --trigger-bucket=test ${deployArgs}`, cwd);
+        const output = run(`${override || cmd} deploy helloData --source=test/test_module/ --trigger-bucket=test ${deployArgs}`, cwd);
         if (override) {
           try {
             assert(output.includes(`/functions/helloData`));
@@ -191,7 +187,7 @@ function makeTests (service, override) {
       });
 
       it(`should deploy a synchronous function`, () => {
-        const output = run(`${override || cmd} deploy helloPromise --local-path=test/test_module/ --trigger-bucket=test ${deployArgs}`, cwd);
+        const output = run(`${override || cmd} deploy helloPromise --source=test/test_module/ --trigger-bucket=test ${deployArgs}`, cwd);
         if (override) {
           try {
             assert(output.includes(`/functions/helloPromise`));
@@ -205,7 +201,7 @@ function makeTests (service, override) {
       });
 
       it(`should deploy a function that throws and process does not crash`, () => {
-        const output = run(`${override || cmd} deploy helloThrow --local-path=test/test_module/ --trigger-bucket=test ${deployArgs}`, cwd);
+        const output = run(`${override || cmd} deploy helloThrow --source=test/test_module/ --trigger-bucket=test ${deployArgs}`, cwd);
         if (override) {
           try {
             assert(output.includes(`/functions/helloThrow`));
@@ -219,7 +215,7 @@ function makeTests (service, override) {
       });
 
       it(`should deploy a function returns JSON`, () => {
-        const output = run(`${override || cmd} deploy helloJSON --local-path=test/test_module/ --trigger-bucket=test ${deployArgs}`, cwd);
+        const output = run(`${override || cmd} deploy helloJSON --source=test/test_module/ --trigger-bucket=test ${deployArgs}`, cwd);
         if (override) {
           try {
             assert(output.includes(`/functions/helloJSON`));
@@ -233,7 +229,7 @@ function makeTests (service, override) {
       });
 
       it(`should deploy a function with a timeout`, () => {
-        const output = run(`${override || cmd} deploy helloSlow --local-path=test/test_module/ --trigger-http --timeout=2s ${deployArgs}`, cwd);
+        const output = run(`${override || cmd} deploy helloSlow --source=test/test_module/ --trigger-http --timeout=2s ${deployArgs}`, cwd);
         if (override) {
           try {
             assert(output.includes(`/functions/helloSlow`));
@@ -247,7 +243,7 @@ function makeTests (service, override) {
       });
 
       it(`should deploy an HTTP function that fails to respond (crashes asynchronously)`, () => {
-        const output = run(`${override || cmd} deploy helloNoResponse --local-path=test/test_module/ --trigger-http ${deployArgs}`, cwd);
+        const output = run(`${override || cmd} deploy helloNoResponse --source=test/test_module/ --trigger-http ${deployArgs}`, cwd);
         if (override) {
           try {
             assert(output.includes(`/functions/helloNoResponse`));
@@ -261,7 +257,7 @@ function makeTests (service, override) {
       });
 
       it(`should deploy an HTTP function`, () => {
-        const output = run(`${override || cmd} deploy helloGET --local-path=test/test_module/ --trigger-http ${deployArgs}`, cwd);
+        const output = run(`${override || cmd} deploy helloGET --source=test/test_module/ --trigger-http ${deployArgs}`, cwd);
         if (override) {
           try {
             assert(output.includes(`/functions/helloGET`));
@@ -275,7 +271,7 @@ function makeTests (service, override) {
       });
 
       it(`should deploy an HTTP function and send it JSON`, () => {
-        const output = run(`${override || cmd} deploy helloPOST --local-path=test/test_module/ --trigger-http ${deployArgs}`, cwd);
+        const output = run(`${override || cmd} deploy helloPOST --source=test/test_module/ --trigger-http ${deployArgs}`, cwd);
         if (override) {
           try {
             assert(output.includes(`/functions/helloPOST`));
@@ -289,7 +285,7 @@ function makeTests (service, override) {
       });
 
       it(`should deploy a function that needs "npm install"`, () => {
-        let _cmd = `${override || cmd} deploy helloUuidNpm --local-path=test/test_module_2/ --trigger-http ${deployArgs}`;
+        let _cmd = `${override || cmd} deploy helloUuidNpm --source=test/test_module_2/ --trigger-http ${deployArgs}`;
         if (!_cmd.includes(`--stage-bucket=${bucketName}`)) {
           _cmd = `${_cmd} --stage-bucket=${bucketName}`;
         }
@@ -307,7 +303,7 @@ function makeTests (service, override) {
       });
 
       it(`should deploy a function that needs "yarn install"`, () => {
-        let _cmd = `${override || cmd} deploy helloUuidYarn --local-path=test/test_module_3/ --trigger-http ${deployArgs}`;
+        let _cmd = `${override || cmd} deploy helloUuidYarn --source=test/test_module_3/ --trigger-http ${deployArgs}`;
         if (!_cmd.includes(`--stage-bucket=${bucketName}`)) {
           _cmd = `${_cmd} --stage-bucket=${bucketName}`;
         }
@@ -325,14 +321,14 @@ function makeTests (service, override) {
       });
 
       it(`should fail when the module does not exist`, () => {
-        const output = run(`${override || cmd} deploy fail --local-path=test/test_module/foo/bar --trigger-http ${deployArgs}`, cwd);
+        const output = run(`${override || cmd} deploy fail --source=test/test_module/foo/bar --trigger-http ${deployArgs}`, cwd);
         assert(output.includes(`Provided directory does not exist.`));
       });
 
       if (!override) {
         it(`should fail when the function does not exist`, () => {
           // TODO: Make this work for the SDK
-          const output = run(`${cmd} deploy doesNotExist --local-path=test/test_module/ --trigger-http ${deployArgs}`, cwd);
+          const output = run(`${cmd} deploy doesNotExist --source=test/test_module/ --trigger-http ${deployArgs}`, cwd);
           assert(output.includes(`Node.js module defined by file index.js is expected to export function named doesNotExist`));
         });
       }
@@ -366,7 +362,7 @@ function makeTests (service, override) {
 
     describe(`list`, () => {
       it(`should list functions`, () => {
-        const output = run(`${override || cmd} list ${override ? `--region=${REGION}` : shortArgs}`, cwd);
+        const output = run(`${override || cmd} list ${override ? `--regions=${REGION}` : shortArgs}`, cwd);
         assert(output.includes(name));
         assert(output.includes(`helloData`));
         assert(output.includes(`helloGET`));
@@ -422,7 +418,7 @@ function makeTests (service, override) {
       });
 
       it(`should re-deploy over a function`, () => {
-        const output = run(`${override || cmd} deploy helloJSON --entry-point=helloPromise --local-path=test/test_module/ --trigger-bucket=test ${deployArgs}`, cwd);
+        const output = run(`${override || cmd} deploy helloJSON --entry-point=helloPromise --source=test/test_module/ --trigger-bucket=test ${deployArgs}`, cwd);
         if (override) {
           assert(output.includes(`/functions/helloJSON`));
         } else {
@@ -536,18 +532,18 @@ function makeTests (service, override) {
 
     describe(`delete`, () => {
       it(`should delete a function`, () => {
-        let output = run(`${override || cmd} delete helloData ${overrideArgs || shortArgs}`, cwd);
+        let output = run(`${override || cmd} delete helloData ${overrideArgs ? overrideArgs + ' -q' : shortArgs}`, cwd);
         if (override) {
           assert(output.includes(`/functions/helloData`));
-          assert(output.includes(`deleted`));
+          assert(output.toLowerCase().includes(`deleted`));
         } else {
           assert(output.includes(`Function helloData deleted.`));
         }
 
-        output = run(`${override || cmd} list ${overrideArgs || shortArgs}`, cwd);
+        output = run(`${override || cmd} list ${override ? '--regions=' + REGION : ''}`, cwd);
         assert.equal(output.includes(`No functions deployed`), false);
         assert.equal(output.includes(`helloData`), false);
-        assert(output.includes(`hello`));
+        assert(output.toLowerCase().includes(`hello`));
       });
     });
 
@@ -631,7 +627,6 @@ function makeTests (service, override) {
         assert(output.includes(`${prefix}`));
         assert(output.includes(`RUNNING`));
         assert(output.includes(`http://localhost:${REST_PORT}/`));
-        assert(output.includes(`http://localhost:${GRPC_PORT}/`));
         assert(output.includes(`http://localhost:${SUPERVISOR_PORT}/${PROJECT_ID}/${REGION}`));
 
         output = run(`${cmd} stop ${shortArgs}`, cwd);
@@ -666,15 +661,6 @@ describe(`system/cli`, () => {
       .then(() => storage.bucket(bucketName).delete());
   });
 
-  if (!process.env.CIRCLECI) {
-    makeTests(`grpc`);
-    makeTests(`rest`);
-    makeTests(`rest`, `${GCLOUD} beta functions`);
-  } else if (process.env.CIRCLE_NODE_INDEX === '0') {
-    makeTests(`grpc`);
-  } else if (process.env.CIRCLE_NODE_INDEX === '1') {
-    makeTests(`rest`);
-  } else if (process.env.CIRCLE_NODE_INDEX === '2') {
-    makeTests(`rest`, `${GCLOUD} beta functions`);
-  }
+  makeTests(`rest`);
+  makeTests(`rest`, `${GCLOUD} beta functions`);
 });