+++ test/unit/model/operation.test.js
@@ -15,27 +15,14 @@
 
 'use strict';
 
-const _ = require(`lodash`);
-const grpc = require(`grpc`);
 const proxyquire = require(`proxyquire`);
 
 describe(`unit/model/operation`, () => {
-  let Operation, mocks;
+  let Operation;
   const TEST_NAME = `operations/abcd1234`;
 
   beforeEach(() => {
-    mocks = {
-      protos: {
-        decode: (arg1) => _.cloneDeep(arg1),
-        decodeAnyType: sinon.stub(),
-        encodeAnyType: sinon.stub()
-      }
-    };
-    sinon.spy(mocks.protos, `decode`);
-
-    Operation = proxyquire(`../../../src/model/operation`, {
-      './protos': mocks.protos
-    });
+    Operation = proxyquire(`../../../src/model/operation`, {});
   });
 
   describe(`Operation`, () => {
@@ -69,12 +56,11 @@ describe(`unit/model/operation`, () => {
           assert(err.message === message);
           assert.errorType(
             err,
-            grpc.status.INVALID_ARGUMENT,
+            3,
             message,
             [
               `DebugInfo`,
-              `BadRequest`,
-              `ResourceInfo`
+              `BadRequest`
             ]
           );
           return true;
@@ -121,24 +107,6 @@ describe(`unit/model/operation`, () => {
       );
     });
 
-    it(`should decode the props`, () => {
-      sinon.spy(Operation, `decode`);
-
-      let props = {};
-      let operation = new Operation(TEST_NAME);
-
-      assert.deepEqual(operation, { name: TEST_NAME });
-      assert(Operation.decode.callCount === 1);
-      assert.deepEqual(Operation.decode.getCall(0).args, [props]);
-
-      props = { done: true };
-      operation = new Operation(TEST_NAME, props);
-
-      assert.deepEqual(operation, _.merge(props, { name: TEST_NAME }));
-      assert(Operation.decode.callCount === 2);
-      assert.deepEqual(Operation.decode.getCall(1).args, [props]);
-    });
-
     it(`should return an Operation instance`, () => {
       assert(new Operation(TEST_NAME) instanceof Operation);
     });
@@ -147,57 +115,6 @@ describe(`unit/model/operation`, () => {
   describe(`Operation.NAME_REG_EXP`, () => {
     it(`should be an instance of RegExp`, () => {
       assert(Operation.NAME_REG_EXP instanceof RegExp);
-    });
-  });
-
-  describe(`Operation.decode`, () => {
-    it(`should decode the props`, () => {
-      let props;
-      let operation = Operation.decode();
-
-      assert.deepEqual(operation, {});
-      assert(mocks.protos.decode.callCount === 1);
-      assert(mocks.protos.decodeAnyType.callCount === 0);
-
-      props = {
-        error: {
-          details: [
-            {}
-          ]
-        }
-      };
-      operation = Operation.decode(props);
-
-      assert.deepEqual(operation, props);
-      assert(operation !== props);
-      assert(mocks.protos.decode.callCount === 2);
-      assert(mocks.protos.decodeAnyType.callCount === 1);
-    });
-
-    it(`should disallow both "error" and "response"`, () => {
-      let props = {
-        error: {},
-        response: {}
-      };
-      assert.throws(
-        () => {
-          Operation.decode(props);
-        },
-        (err) => {
-          const message = `Operation may only have one of 'error' or 'response'!`;
-          assert(err instanceof Error);
-          assert(err.message === message);
-          assert.errorType(
-            err,
-            grpc.status.INVALID_ARGUMENT,
-            message,
-            [
-              `BadRequest`
-            ]
-          );
-          return true;
-        }
-      );
     });
   });
 
@@ -215,82 +132,5 @@ describe(`unit/model/operation`, () => {
 
   describe(`Operation.parseName`, () => {
     it(`should parse a formatted Operation name string`);
-  });
-
-  describe(`Operation#toProtobuf`, () => {
-    it(`should return a representation suitable for serialization to a protobuf`, () => {
-      sinon.spy(Operation, `decode`);
-
-      const message = `error`;
-      const operation = new Operation(TEST_NAME, {
-        metadata: {},
-        error: {
-          message,
-          code: 1,
-          details: [
-            {}
-          ]
-        }
-      });
-
-      assert(Operation.decode.callCount === 1);
-
-      const proto = operation.toProtobuf();
-
-      assert.deepEqual(proto, operation);
-      assert(Operation.decode.callCount === 2);
-      assert(mocks.protos.encodeAnyType.callCount === 3);
-
-      operation.error = new Error(message);
-      operation.error.code = 1;
-      operation.error.details = [
-        {}
-      ];
-
-      const proto2 = operation.toProtobuf();
-
-      assert.deepEqual(proto2, proto);
-      assert(Operation.decode.callCount === 3);
-      assert(mocks.protos.encodeAnyType.callCount === 6);
-
-      delete operation.metadata;
-      delete operation.error;
-
-      const proto3 = operation.toProtobuf();
-
-      assert.deepEqual(proto3, { name: TEST_NAME });
-      assert(Operation.decode.callCount === 4);
-      assert(mocks.protos.encodeAnyType.callCount === 8);
-
-      operation.error = new Error(message);
-      operation.error.code = 1;
-
-      const proto4 = operation.toProtobuf();
-
-      assert.deepEqual(proto4, {
-        name: TEST_NAME,
-        error: {
-          code: 1,
-          message
-        }
-      });
-      assert(Operation.decode.callCount === 5);
-      assert(mocks.protos.encodeAnyType.callCount === 10);
-    });
-  });
-
-  describe(`Operation#toJSON`, () => {
-    it(`should return a representation suitable for serialization to JSON`, () => {
-      sinon.spy(Operation, `decode`);
-
-      const operation = new Operation(TEST_NAME);
-
-      assert(Operation.decode.callCount === 1);
-
-      const json = operation.toJSON();
-
-      assert.deepEqual(json, operation);
-      assert(Operation.decode.callCount === 2);
-    });
   });
 });